# mock-server/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Clinic OS API Specification (SCS)
  description: |
    المستند البرمجي لواجهات برمجة التطبيقات (API Specification) لمنصة عيادتي الذكية.
    يحتوي الملف على مسارين أساسيين:
    1. **واجهات الأوبريشن والمشغل (Platform Operations):** تبدأ بـ `/admin/v1` ومحمية بـ VPN.
    2. **واجهات العيادات والعملاء (Clinic Dashboard):** تبدأ بـ `/v1` وتعتمد على ترويسة العيادة.
  version: 1.0.0
servers:
  - url: https://api.SCS-admin.com
    description: Public Clinic Dashboard Gateway
  - url: https://api.SCS-ops.com
    description: Private Operations Gateway (Accessible via VPN only)
  - url: http://localhost:3000
    description: Local Mock Development Server
paths:
...
  /admin/v1/auth/login:
    post:
      summary: تسجيل دخول مدير المنصة (Platform Admin Login)
      description: تسجيل الدخول للوحة تحكم الأوبريشن باستخدام البريد الإلكتروني وكلمة المرور.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: ops@SCS-ops.com
                password:
                  type: string
                  example: SecurePassword123!
      responses:
        '200':
          description: نجاح طلب الدخول وتوليد كود التحقق الثنائي (2FA)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      two_factor_required:
                        type: boolean
                        example: true
                      temp_token:
                        type: string
                        example: temp_jwt_token_for_2fa
  /admin/v1/auth/verify-2fa:
    post:
      summary: التحقق من كود التحقق الثنائي (Verify 2FA)
      description: إكمال عملية تسجيل الدخول للأدمن باستخدام كود OTP المرسل على الهاتف.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - temp_token
                - otp_code
              properties:
                temp_token:
                  type: string
                  example: temp_jwt_token_for_2fa
                otp_code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: تسجيل دخول ناجح وإصدار الـ Tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      admin:
                        type: object
                        properties:
                          id:
                            type: string
                            example: admin-uuid-12345
                          email:
                            type: string
                            example: ops@SCS-ops.com
                          full_name:
                            type: string
                            example: أحمد مشغل النظام
                          role:
                            type: string
                            example: super_admin
                      access_token:
                        type: string
                        example: jwt_access_token_here
                      refresh_token:
                        type: string
                        example: jwt_refresh_token_here
...